refactor: use absolute imports for cogs
