+++ ironaccord_bot/cogs/gm.py
@@ -2,9 +2,9 @@ import discord
 from discord import app_commands
 from discord.ext import commands
 
-from ..models import database as db
-from ..utils.embed import simple
-from ..ai.mixtral_agent import MixtralAgent
+from ironaccord_bot.models import database as db
+from ironaccord_bot.utils.embed import simple
+from ironaccord_bot.ai.mixtral_agent import MixtralAgent
 
 class GmCog(commands.Cog):
     def __init__(self, bot: commands.Bot):
