+++ ironaccord_bot/cogs/help.py
@@ -2,7 +2,7 @@ import discord
 from discord import app_commands
 from discord.ext import commands
 
-from ..utils.embed import simple
+from ironaccord_bot.utils.embed import simple
 
 class HelpCog(commands.Cog):
     def __init__(self, bot: commands.Bot):
