+++ ironaccord_bot/cogs/mission.py
@@ -5,9 +5,9 @@ import discord
 from discord import app_commands
 from discord.ext import commands
 
-from ..utils.embed import simple
-from ..utils.mission_engine import resolve_choice
-from ..models import mission_service
+from ironaccord_bot.utils.embed import simple
+from ironaccord_bot.utils.mission_engine import resolve_choice
+from ironaccord_bot.models import mission_service
 
 MISSIONS_PATH = Path(__file__).parent.parent / 'data' / 'missions'
