+++ python_bot/tests/conftest.py
@@ -2,6 +2,6 @@ import sys
 from pathlib import Path
 
 # Ensure the project root is on the path so `import python_bot` works
-ROOT = Path(__file__).resolve().parents[1]
+ROOT = Path(__file__).resolve().parents[2]
 if str(ROOT) not in sys.path:
     sys.path.insert(0, str(ROOT))
